Minor Changes
Made the comment a little less snarky.
Forced the bot to fetch new posts after it checks the 100 it grabbed. (It will recheck posts which is good as it monitors conversation)
Added a status message that gets sent to me every hour.
Added a switch so the bot goes back and forth between checking 100 new posts and 100 hot posts.